Move paint and event dispatching into separate file
